주문하기 실행 시 재고 감소를 비동기처리로 구현

+++ src/main/java/com/avengers/musinsa/domain/product/entity/ProductVariant.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class ProductVariant {
-    private Integer productVariantId;
+    private Long productVariantId;
 
     private Product product;
     private Long productId;
